# Conflicts:
#	src/main/java/at/qe/skeleton/services/BookmarkService.java
#	src/main/java/at/qe/skeleton/services/BorrowService.java
#	src/main/java/at/qe/skeleton/services/UserService.java
#	src/test/java/at/qe/skeleton/tests/MediaServiceTest.java

+++ src/main/java/at/qe/skeleton/ui/controllers/UserDetailController.java
@@ -1,11 +1,15 @@
 package at.qe.skeleton.ui.controllers;
 
 import java.io.Serializable;
+import java.util.Collection;
 import java.util.List;
 
 import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
 
+import at.qe.skeleton.model.Borrowed;
+import at.qe.skeleton.model.Reserved;
+import at.qe.skeleton.services.BorrowService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
@@ -31,6 +35,14 @@ public class UserDetailController implements Serializable {
 
 	@Autowired
 	private UndoRedoService undoRedoService;
+
+	@Autowired
+	FMSpamController fms;
+
+	@Autowired
+	BorrowService borrowService;
+
+
 
 	/**
 	 * Attribute to cache the currently displayed user
@@ -82,6 +94,8 @@ public class UserDetailController implements Serializable {
 			// TODO: Exception-Handling
 		}
 		this.doReloadUser();
+		fms.info("A new user with the username: "+ username +" was created!");
+		fms.info("Please reload the page.");
 	}
 
 	/**
@@ -98,30 +112,36 @@ public class UserDetailController implements Serializable {
 		this.undoRedoService.addAction(undoRedoService.createAction(userService.loadUser(this.user.getUsername()), user,
 				UndoRedoService.ActionType.EDIT_USER));
 		this.user = this.userService.saveUser(this.user);
-
-		FacesMessage asGrowl = new FacesMessage(FacesMessage.SEVERITY_INFO, "Changes saved!", "");
-		FacesContext.getCurrentInstance().addMessage("asGrowl", asGrowl);
+		fms.info("Changes saved");
 	}
 
 	/**
 	 * Action to delete the currently displayed user.
 	 */
 	public void doDeleteUser() {
-		try {
-			this.userService.deleteUser(this.user);
-			this.undoRedoService.addAction(undoRedoService.createAction(user, UndoRedoService.ActionType.DELETE_USER));
 
-			FacesMessage asGrowl = new FacesMessage(FacesMessage.SEVERITY_WARN, "User was deleted!", "");
-			FacesContext.getCurrentInstance().addMessage("asGrowl", asGrowl);
-		} catch (UserService.UnauthorizedActionException unauthorizedActionException) {
-			System.out.println(unauthorizedActionException.getMessage());
-			// TODO: Exception-Handling
+		Collection<Borrowed> d1 = borrowService.getAllBorrowsByUser(this.user);
+		if (d1.size() > 0) {
+				fms.warn("This user has still borrowed: " + d1.size() + " article(s) and cannot be deleted yet.");
+		} else {
+				try {
+					this.userService.deleteUser(this.user);
+					this.undoRedoService.addAction(undoRedoService.createAction(user, UndoRedoService.ActionType.DELETE_USER));
+
+				} catch (UserService.UnauthorizedActionException unauthorizedActionException) {
+					System.out.println(unauthorizedActionException.getMessage());
+
+				}
+				this.user = null;
+				fms.info("The user was deleted and all his bookmarks and reserved media has been deleted");
+				fms.info("Please reload the page.");
 		}
-		this.user = null;
+
 	}
 
 	public void changeUserRoles() {
 		userService.changeUserRoles(user, newRolesString);
+		fms.info("Role edit done successfully.");
 	}
 
 	public UserService getUserService() {
